// Web app/csi-mkd-premarital-app-FE/src/api/api-main-app/models/app-feedback-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AppFeedbackResponseDto {
  createdAt?: string;
  id?: string;
  improvements?: string | null;
  likedMost?: string | null;
  locale?: string | null;
  page?: string | null;
  rating?: number;
}